requirements.txt
django==1.10.7
#gunicorn==19.3.0
#cymysql==0.8.1
#django-cymysql==0.3.1
pymongo==3.4.0
django-multiselectfield=0.1.5
psycopg2==2.7.1
django-bitfield==1.9.3
